This is a fix for an off-by-one error that causes the first base of the stop codon to appear in a gene's CDS sequence in some circumstances. This is a known bug that came up again recently on the Apollo mailing list.  I posted the following to the list but never received any feedback. I'm fairly confident this is correct, so I'm comitting the fix:
Hi Rob-

I think this bug may be simpler than I had originally thought.  Although
I can't find this explicitly documented anywhere in the main Apollo
codebase, the following comment (in ChadoCds.java, part of the chado
JDBC adapter) leads me to think that FeatureSet.translation_end is also
supposed to be set to the (genomic) location of the first base in the
stop codon, and not the last base of the CDS.  Mark, does this sound
right to you?  I don't *think* I wrote the first part of the following
comment:

  /** Return basepair where translation stop codon begins. 2 less than
high on
      forward strand, 2 more than low on rev strand. This assumes that the
      cds contains the stop, which it should. if it doesnt the coords
should be
      adjusted to do so. fb avoids this by doing automatic stop.  */
  // JC: Note that this will break if the stop codon spans multiple exons
  int getTranslationEnd() {
    if (isForward())
      return getEnd() - 2;
    else
      return getEnd() + 2;
  }

So if that's the case then the problem seems clear, namely that in
FeatureSet.get_ORF(String mRNA) we need to subtract 1 from the (edited
mRNA coordinate of) the translation_end position in order to not include
any of the stop codon in the CDS.  Note that I'm leaving aside for the
moment the other option, which is to change the CDS of *both* the
transcript and the (final coding) exon to include the stop codon,
GenBank-style.  So in FeatureSet.java I would suggest changing this:

        int end_offset = ((translation_end > 0) ?
                          getFeaturePosition(translation_end) : -1);

to this:

        int end_offset = ((translation_end > 0) ?
                          getFeaturePosition(translation_end) - 1: -1);

And, of course, adding some documentation that makes it clear why the -1
is there and what translation_end is supposed to be set to.  This
modification appears to correct the bug for me, but I'll hold off on
committing anything until I've done some more testing and others have
had a chance to weigh in on the issue.

Jonathan

p.s. If my supposition about translation_end is correct, I might suggest
changing the variable's name to "stop_start"  :)