Fixed a rather pernicious bug that I probably introduced in the first place.  Note that "results" and "annotations" were previously being created anew directly within TigrChadoInstance, instead of via calls to getResultStrandedFeatSet and getAnnotStrandedFeatSet.  I think this was either as a result of a refactoring in FlybaseChadoInstance that didn't get carried over to TigrChadoInstance, or maybe because I didn't pay attention to the existing code when implementing the latter.
Anyway, the problem occurred when I started using addOneLevelResults();
the features being created in that method weren't showing up in Apollo.
The reason is because of this call:

    StrandedFeatureSet parentFeatSet = ci.getResultStrandedFeatSet();

This call was creating a *new* stranded feature set not connected to the
one being initialized in TigrChadoInstance (because it wasn't using
getResultStrandedFeatureSet itself.)  So the immediate problem has
been fixed, but the real issue here is that some of the methods in
JdbcChadoAdapter (e.g., addGeneModels) take a StrandedFeatureSet as
an explicit parameter, whereas other methods in JdbcChadoAdapter
(e.g., addOneLevelResults) obtain the StrandedFeatureSet via the "back
door", by using ChadoInstance.getResultStrandedFeatSet.  I guess I'd
be in favor of going with the former approach (using an explicit
parameter) for all the methods.