Added a global config option, "SaveClonedFeaturesInTransactions", set to false by default.  The various apollo Transaction classes always store a pointer/reference to the affected SeqFeatureI.  When this new config. option is set to true, these transaction classes will *also* store a cloned copy of the affected SeqFeatureI (which can be accessed by a new method in each of the modified classes.)
The reason for this is that that Apollo transaction objects currently
store a mix of mutable and immutable objects.  Therefore, by extension,
the transactions themselves are mutable.  This fact does not appear to
be well-documented (although maybe it should be obvious...) and I would
argue that it makes the transactions difficult to work with.  Take, for
example, the AddTransaction that is created when a new gene annotation
is created.  This transaction stores a pointer to the added gene
(SeqFeatureI), which I'll call G1.  Let's say that initially G1 has a
single transcript, T1.  So after the gene is added we have the following
in our list of uncommitted transactions:

AddTransaction (added feature = G1[transcript list = T1]) at time t1

Now let's say that the user subsequently changes his mind and decides
to delete G1.  As part of the delete T1 will be removed from G1's
transcript list.  As a result the transaction that appears above now
looks like this (since G1 has been updated):

AddTransaction (added feature = G1[transcript list = empty]) at time t1

This makes it look like a gene with no transcripts was added at time
T1, which is not the case (or at least it does not reflect the original
"official" version of events stored in the transaction manager.)

Of course it is possible to work with these transactions, but in order
to interpret them correctly it's necessary for the code that queries
the transactions to know how "deep" it's allowed to go into the data
structure.  It seems as though this is something that should be
encapsulated by (or at least documented in) the transaction classes
themselves.  It is also possible to work around this particular problem
by using the TransactionManager's coalesce() method to eliminate both
the add and delete transactions completely.  However, I think the
criticism of the original transaction list still stands, and the coalesce
function isn't something that should be used to correct inconsistencies
in the transaction list.

In any case, to work around this problem while maintaining backwards
compatibility I've added the aforementioned global config. option, which
*must* be set to true whenever the PureJDBCTransactionAdapter is in use
(this needs to be improved.)  When set to true all the Transaction
objects will automatically clone their primary SeqFeatureI, and this
cloned copy can be retrieved by calling the appropriate method (e.g.,
AddTransaction.getAddedFeatureClone()).  I didn't want to simply replace
the original reference with a reference to a clone because it's entirely
possible that the existing code *relies* on the fact that the SeqFeature
objects are mutable.  I don't know.  Therefore the safest thing to do
was *add* a new field for the clone, and provide a config. option to
minimize the amount of extra memory taken (only space for the extra
reference is required if the option is set to false.)