Changes in this commit (sorry -- these should really have been committed separately, but since they're not all easily divided into sets of source files I'm lumping them together):
1. Added several new configuration parameters to chado-adapter.xml:

 PureJDBCCopyOnWrite
  -set to true to enable copy-on-write (not yet implemented; details
   to follow) in PureJDBCTransactionWriter
 PureJDBCNoCommit
  -set to true to force the PureJDBCTransactionWriter to rollback all
   commits, even when otherwise successful (used for debugging)
 transGeneRelationTerm
  -cvterm.name of the relationship between transcript and gene
 exonTransRelationTerm
  -cvterm.name of the relationship between exon and transcript
 cdsTransRelationTerm
  -cvterm.name of the relationship between CDS and transcript
 polypeptideCdsRelationTerm
  -cvterm.name of the relationship between polypeptide and CDS
 polypeptideTransRelationTerm
  -cvterm.name of the relationship between polypeptide and transcript

The various relation parameters are used by JdbcChadoAdapter and
PureJDBCTransactionWriter, instead of having the names hard-coded.
Note that the last one (polypeptide-transcript) is a redundant
relationship used by the TIGR annotation databases.  It is set to null
by default, meaning that this relationship will not be read or written
by the adapter.

Also note that the names of the various cvterm properties still need to
be standardized.

2. Modified the chado adapter to read and write the chado
featureloc.is_fmin_partial and featureloc.is_fmax_partial columns.
These are used for CDS features in particular when the coding sequence
runs off one (or both) ends of a transcript.  Apollo stores such
out-of-range values internally as "-1" coordinates, but in the
database they are represented by using the appropriate end coordinate
of the transcript plus the is_(fmin|fmax)_partial flag.

3. Changed signatures of some of the update methods in JdbcChadoAdapter.java

4. Improved/added delete code in JdbcChadoAdapter and PureJDBCTransactionWriter.