Modifications to allow the PureJDBCTransactionWriter to handle MERGE and SPLIT CompoundTransactions (i.e., merging and splitting transcripts and genes in Apollo.) Some workarounds were needed to get everything working, but I believe they should all be reverse compatible.  (It is difficult to be sure, however, since some of the logic in AnnotationEditor is quite convoluted.)  From the point of view of the PureJDBCTransactionWriter there are two main problems with the transactions generated by the AnnotationEditor, particularly in the case of SPLIT events:
#1: Transactions are generated for features that have not yet been assigned
    an id (SeqFeatureI.setId()) and a name (SeqFeatureI.setName()).
#2: No Transaction is generated for certain (important) events.

#1 is a problem only because of the code that I had added previously that
(optionally) clones SeqFeatures (and provides methods to return those clones)
in the various Transaction classes.  If you recall my goal was to have all the
Apollo Transactions record the following information:

 -the affected features, as they were *before* any change was applied
 -the change that was applied

The idea was that given some initial state and a set of transactions, one should
be able to reach the correct final state by simply applying each transaction in
turn.  This was not easy (or perhaps even possible) the way Apollo was set up
originally, because the Transactions stored pointers to mutable feature objects,
so what each Transaction actually recorded was the following:

 -the affected features, after *all* the changes have been applied
 -the change that was applied (although not always, since in some cases the
  mutable object itself was used to encode the new value)

In any case, the changes I made to address this problem--primarily adding the
option to clone features in Transactions--in conjunction with a couple of other
factors (namely the fact that the TIGR name adapter doesn't generate temp ids
and also needs to work without *requiring* the transaction coalescer), mean that
all SeqFeatures need to be assigned an ID before participating in any other
types of Transactions.  SPLIT Transactions in particular are problematic because
the AnnotationEditor generates a new gene and transcript and then shuffles the
existing exons around, generating an Add event followed by a bunch of
UpdateParentTransactions.  Because of the way this was implemented the names
and ids didn't get set/updated until after the UpdateParentTransactions had
been processed.  To fix this I reworked some of the code in AnnotationEditor.
In particular I replaced splitGeneSetup() with setSplitGeneIds().  As the name
suggests, the new method *only* sets the ids and names and performs the other
changes elsewhere.  This improved things but didn't completely fix the problem,
as the dependencies (between the parent changes and the id setting) are somewhat
complicated.  To address this I added a method to Trasaction and ChangeList
called updateClonedTranscriptIdsAndNames.  This method traverses all the
Transactions in a ChangeList and--if cloned features are being stored--updates
each null (or "no_name") id or name in each cloned feature using the corresponding
id and name of the current feature.  This allows some flexibility in the order
the transactions are created while still allowing them to accurately reflect
what (in principle) occurred.

The main issue for #2 also comes up during SPLIT transactions, where the
AnnotationEditor generates a new Transcript without firing/generating a
corresponding AddTransaction.  It does, however, generate UpdateParentTransactions
involving this Transcript, which means that any code trying to follow the data
model changes using the transactions is going to be confused, because it will
start seeing UpdateParentTransactions for this Transcript without ever knowing
that the Transcript was added.  This is fine if the adapter uses the coalescer,
because the Transcript is only temporary, but it's problematic if you're trying
to do what the PureJDBCTransactionWriter does.  As a workaround I've modified
the PureJDBCTransactionWriter to be a little more forgiving about how it processes
updates.  If, for example, it comes across an UpdateParentTransaction where the
first half (the delete) has already been processed, but the second half (the
add) has not, it apply only the second half to the database.  Similarly, if the
entire transaction has already been processed, it is ignored.  This can happen
because the PureJDBCTransaction writer interprets Apollo gene AddTransactions as
meaning "add this gene, its transcripts, and any attached exons" (as a workaround
for the fact that Apollo will not generate an add event for any transcript.)
Therefore it may see UpdateParentTransactions for exons that have already been
added or moved to a particular gene.

Other modifications/notes:
  -Added support for NAME and ID update transactions in PureJDBCTransactionWriter.
  -Commented out some code that was generating duplicate NAME change events
  -The calls to setNewSplitFeature() don't make any sense to me.  Since there
   don't seem to be any calls to getNewSplitFeature() anywhere in the code,
   perhaps these calls (and the instance variable) can be removed completely?
  -UpdateParentTransaction now takes the parent as an explicit argument,
   instead of trying to pull it from the feature itself.