Added two new parameters that give some control over how feature_id lookups are performed by the getFeatureId() method of JdbcChadoAdapter:
queryFeatureIdWithUniquename
 -include chado feature.uniquename in the query for a feature's id
queryFeatureIdWithName
 -include chado feature.name in the query for a feature's id

If both are set to true then an SQL "OR" clause will be used (the
default, for reverse compatibility.)  If both are set to false
then the adapter will complain and use just the uniquename.  Some
sites may have an index on uniquename but not name and may know
that they only need to do the lookup on uniquename.  In such cases
it's useful to be able to set these parameters accordingly.