There are two changes in this commit:
1. Refactored handling of boolean values.  Previously some were hard-coded
and some used the (DBMS-dependent) methods getBooleanTrueValue() and
getBooleanFalseValue(), which returned String values.  These two methods
have been replaced by a single method, getBooleanValue(boolean val),
which returns an Object.  The PostgreSQL adapter class returns Booleans
and the Sybase adapter class returns Integers (in Sybase we use a BIT or
TINYINT datatype, as there's no boolean type.)  The useful thing about
this approach (returning Objects) is that the Objects can be passed
directly to the update methods.  On the read/query side, calling toString()
on any of these Objects will give a string literal that can be substituted
directly into an SQL query (since postgres is happy with either 'true', 't',
or true).  This code should work with both PostgreSQL and Sybase, although
I'm still checking through the other classes to see if there are datatype-
dependent function calls (e.g., in ChadoFeatureLoc and the like.)

2. Modified 'owner' and 'date' featureprop code so that if the featureprop
is not present then a warning is printed but the commit/save does not
fail.  Olivier, if you'd like to have the save operation fail in this case
let me know and I'll add a parameter that controls how severe the error
should be.