Fixed a couple of bugs and moved more cvterms into the XML configuration file.  Previously the cvterm name for target sequence descriptions (either "description" or "defline") was hard-coded in the site-specific ChadoInstance files and passed as an argument to addSearchHits.  It is now specified by the following two (new) parameters, with the defaults set to maintain reverse compatibility:
 seqDescriptionCVName - the same as propertyCVName by default
 seqDescriptionTerm - "description" by default

(We're gradually working our way towards eliminating the site and
chado-version specific Java classes and replacing them with XML
configuration options!)

A recent commit means that the PureJDBCTransactionWriter now inserts
"owner" and "date" featureprops with each newly-added feature.  Since
I don't *think* there's universal agreement on these terms in chado
I've made them configurable, with the following new parameters (again
with the defaults set to leave the present behavior unchanged).  The
query code has also been modified to use these terms where
appropriate:

 featureOwnerPropertyTerm - "owner" by default
 featureCreateDatePropertyTerm - "date" by default

There was also a slight problem with the featureprop code, namely
that featureprop.rank is NOT nullable.  The insert code has been
modified to take this into account and use rank=0, since there
should only ever be 1 of each of these properties.

In XmlConfigFileParser there was a bug with the way that boolean
properties were handled, specifically in the presence of an
<inheritsInstance> tag.  The basic problem is that I was using
"boolean" instead of "Boolean" and was defining default (boolean)
values in the parser.  Therefore if you had a boolean property with a
default of false that was set to true in the "super-instance"--but not
defined in the child instance--it would end up being set to the
default value, not the parent value.  The fix for this was to convert
all the booleans to Booleans and only set the property if the Boolean
is != null (as we're doing for most of the non-boolean props.)  This
change eliminates another problem, which is that many of the default
(boolean) values were defined *both* in the XmlConfigParser and also
in the AbstractChadoInstance.  Now the defaults should be defined in
the latter but not the former.

Also did some cleaning up, mostly rearranging methods and variables so
that similar properties (e.g., relationship cvterms) are grouped
together a little better.  At some point I'd still like to impose a
uniform naming scheme on all the property names; they're mostly OK but
a handful aren't consistent.