Fixed remaining bugs (hopefully) in SPLIT transactions.  Two main issues were addressed:
1. In some cases the actual and declared order of events was different.
In particular we had NAME and ID update transactions in which the ID is
updated first but the NAME update transaction still had a (cloned)
reference to the old id.  So the PureJDBCTransactionWriter was doing
something like this:

 - update ID of object from A1 to A2
 - try to find object A1 to update its NAME ( -> fails)

To fix this the ids and names stored in the cloned feature were tweaked
to make it look like the updates were done sequentially.

2. No update ID or update NAME events were being generated for the
old transcripts in a SPLIT event (although the code *was* updating the
names and ids of these objects in Apollo itself.)  Added transactions
for these events.

Currently working on some similar fixes for certain MERGE operations.