Bug fixes for TIGR name adapter.  The problem in this case stemmed from the method being used to determine whether a given feature had already been assigned a name or id by the name adapter (wrt the current Apollo annotation session.)  The code was checking whether the id or name was a valid chado uniquename (according to the TIGR conventions) and, if so, taking that as prima facie evidence that the id was newly-assigned. However, this method breaks when the user tries to create a new gene based on an existing gene prediction (versus a search result) since the way the TIGR dbs are configured the gene predictions will already have valid chado uniquenames in Apollo, but those names won't have been generated by the name adapter.  So, for example, the user drags a transcript containing eha3.exon.1234.1 into the annotation area.  Apollo creates a new transcript and exon but assigns that exon the same id as the exon from which it is derived - eha3.exon.1234.1.  The name adapter is then called to generate a new id, but--seeing that eha3.exon.1234.1 is already a well-formed uniquename--returns that instead of generating a new name.
The current fix (which isn't ideal) is to have the TigrAnnotNameAdapter
explicitly track the ids and names it has assigned, rather than trying
to infer this indirectly.  Some additional modifications were made to
the logic to ensure that exons (generateExonId) get special treatment;
by default a new id or name will not be generated for an exon that has
already been assigned one.  This is necessary because the AnnotationEditor
has a habit of calling generateExonId twice on the same exon (with the
same gene name.)  For other feature types, on the other hand, multiple
calls to generateName will create multiple distinct ids.  This is
needed to support the merge transcripts code, which needs to generate a
new id/name in case the user wants to assign a *new* name to the merged
gene.