Added/improved support for DbXrefs.  On the read side of the equation the JdbcChadoAdapter now sets the "isPrimary()" and "isSecondary()" properties of the DbXrefs correctly.  If there is a dbxref that is both a primary and a secondary dbxref (i.e., it's referenced from both feature.dbxref_id and feature_dbxref.dbxref_id for the same feature) then only one DbXref object will be created, with both isPrimary() and isSecondary() returning true.
On the write side of the equation (no pun intended), the
PureJDBCTransactionWriter now handles AddTransactions with subpart =
DBXREF.  These are generated when genes with dbxrefs are merged.
The one limitation of the current implementation is that it assumes
a request to add a secondary dbxref will always specify a new dbxref
(i.e., it doesn't check the feature_dbxref table to see if such a
link is already present.)  On reflection this probably needs to be
fixed since--particularly in the case of merged genes--it is
conceivable that the two genes might already have a reference to
the same external accession.  This will be fixed in a subsequent
commit.