Added two new methods to ApolloNameAdapterI, generateNewId() and generateNewExonId().  These methods are straightforward variants of generateId() and generateExonId().  In fact, the only difference is that a call to one of the "New" methods should be construed as a _request_ to the name adapter to generate a new id (i.e. an id that isn't assigned to any feature currently in Apollo--including the feature that's passed to the method.)  The name adapter is free to ignore this request if desired, in which case a call to generateNewId() will become a call to generateId() and a call to generateNewExonId() will become a call to generateExonId().  The DefaultNameAdapter and ParameciumNameAdapter have both been modified to ignore the request in this manner, so their behavior will not change.  Only the TigrAnnotNameAdapter currently pays attention to this new piece of information.
I hesitated to make this change, but didn't see any other easy way
to get the information I needed in the name adapter.  The problem is
that the ApolloNameAdapterI interface doesn't really say anything
about whether the ids generated by these methods should be newly-
generated.  Looking at the code it is clear that in some situations
the ids and/or names _will_ be newly-generated, but in other cases
multiple calls to the same method with the same arguments will generate
the same response (e.g., certain cases of the transcript naming code
in the GmodNameAdapter.)  This lack of specificity is at odds with
the AnnotationEditor, which in certain situations really needs to be
able to generate *new* ids.  One example of this is in the merge
gene dialog, where a generateId() call is used to generate a proposed
3rd gene name that the user may choose.  If the method returned one
of the existing ids then the dialog presented to the user wouldn't
have 3 distinct choices.  Another example is in operations that
entail cloning existing features e.g., creating a new gene model
from an existing gene prediction, or duplicating a transcript.  In
these cases the new (cloned) features are created with the same id
and name as the template features, and then a call to generateId is
used to assign the new feature a different name.

Note that this change is reverse compatible, at least for anyone whose
name adapter extends DefaultNameAdapter, or is currently in CVS.  Also
I think I can reasonably argue that it doesn't make the name generation
logic any more convoluted than it already is (perhaps this is not
such a good selling point!)  Ultimately I'd like to clean up the API
so that the temporary/non-temporary id distinction is clear and also
this question of whether an id generation request is allowed to return
with an id that has already been assigned.