Small thing first:  Changed FeatureTreeFrame so that when a new set is loaded it no  longer hides, but instead just updates its set. Should save the  user from having to redisplay it.
Bigger thing:
 I've added the ability to save/load the layout of the ApolloFrame
to a file. The motivation behind this was to make it easier to
setup the display of multispecies synteny data. However I think it
will have several other possible uses. The layout is saved using
the 'Save window layout' item in the File menu and loaded using the
'Load window layout'.

The following things are saved:
  - Size and location of ApolloFrame
  - Location and Orientation of splitpane divider between
    EvidencePanelContainer and SyntenyPanel
  - Location of splitpane divider in each EvidencePanel
  - Location of splitpane dividers in SyntenyPanel
  - For each StrandedZoomableApolloPanel
      - control panel and nav bar visibility
      - Location of splitter view dividers
      - Forward strand visibility
      - Reverse strand visibility
      - Results visibility
      - Annotation visibility
      - Scale visibility
      - Whether sites are shown when zoomed in
      - Whether edge matches are highlighted
      - Whether display is reverse complemented
      - Centre base *
      - Zoom factor *

The base position specific data (*'d) is only applied if the
'Update base location' checkbox is ticked in the 'Load window layout'
JFileChooser dialog.

I've also added a Style variable 'LayoutIniFile' which specifies a
file to use to set the window layout when a set using that style is
loaded. Base position data is not loaded.

One check that is done to make sure the layout data is compatible
with the current displayed data is to check that the number of species
is the same. The ini file also has a version number stored which is
checked. If base position data is being loaded then the range of the
current curation set is checked against the range that was present
when the layout was saved and a warning given if they differ.