Various improvements to ensj adapter:   Sub classed FeatureSetBuilder so I can group on Name rather than Id for homols. That gives me the freedom to use Id for the internal database id of each feature rather than having to set it to the same thing for all members of a set.
 Added ability to load Evidence (ensembl has supporting features in the
db which can be loaded as evidence).

 Added option to add a prefix to the type names of genes (for when loading
from multiple dbs using the builder adapter when each db has the same types
but you want separate tiers).

 Fixed a bug where TypePanel would ignore changes to the included types
in the shown types list when the adapter was brought up with the list
open (happened with BuilderAdapter on second load)

 Work around class for JComboBox not drawing its popup correctly (similar
to menu popup work around necessary in gui).

 System.outs to log4j