Changes to make evidence panel more useful.   It now saves the widths for each feature type when they are modified by the user. Having to resize them every time I switched between types drove me crazy. These widths are stored in the FeatureProperty for this type so that they are propogated to the other evidence panels in multi species view and can also be saved by the user into the type settings (tiers) file.
  It also saves the sort column selected by the user into the
FeatureProperty.

  Tooltips appear over column headings which display the heading
and cells which display the contents of the cell. This was to get round
the problem that the cells are often too small to display the complete
contents. Now the tooltip will show them.

  The EvidencePanel now also listens for PropertySchemeChangeEvents to
allow propogation of changes made to the columns by the TableColumnChooser
Implementing this also fixed a bug where changing colour in the TypesPanel
didn't propogate to the currently selected data in the EvidencePanel