Now it deletes a map when it's last subsection has been moved from it.
The location of the "move subsection" now follows the mouse instead of the middle of the subsection map (which was causing problems).

Fixed Bug where the undo map move subsection code was removing the map data of the map it was trying to display.  This only caused problems when it was a temporary map id that was being used (otherwise the code would just hit the databasea again).

Fixed an issue when adding a subsection to the very back of a map where the previous ending subsection was being pushed back on the map.

Also, fixed a typo in AppInterface "|" instead of "||".