There was a significant bug in how CMap was storing saved links in Oracle.  The issue was caused because a binary object was being stored in a text field (type clob).
The solution is to change the column type of the session_step_object column in the cmap_saved_link table from clob to blob.

DBI::Oracle requires special handling when inserting to type blob so the methods that inserted/updated saved_links needed placed in the Oracle specific module.

Also, added the upgrade doc for 0.16 to 1.0 so that I could add a disclaimer to the Oracle users.